Time: 11 ms (62.06%), Space: 22.6 MB (72.09%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -17,7 +17,7 @@ public:
         q.push(root);
         while(!q.empty()){
             int size = q.size();
-            long long int sum = 0;
+            long sum = 0;
             for(int i = 0;i<size;i++){
                 TreeNode* temp = q.front();
                 q.pop();
